fix : Litigation 진행률 수정 및 조회 버그 수정

+++ src/main/java/GoEasy/Pansori/dto/litigation/LitSimpleDto.java
@@ -33,7 +33,7 @@ public class LitSimpleDto {
         Integer totalStep = 5;
         Long curStep = litigation.getStep() + 1;
 
-        return (curStep / totalStep) * 100;
+        return (float)(curStep) / (float)(totalStep) * 100;
     }
 
 }
